remove lgtm alert

+++ src/utils/keymap.c
@@ -7,7 +7,7 @@ void keymap_init(devices_info *dev_info, unsigned short line, unsigned short col
     dev_info->mapping.line = line;
     dev_info->mapping.colonne = colonne;
     dev_info->mapping.offset = offset;
-    dev_info->mapping.keys = calloc(line * colonne, sizeof(key));
+    dev_info->mapping.keys = calloc((size_t)(line * colonne), sizeof(key));
 
     if (dev_info->mapping.keys == NULL)
     {
